pyproject.toml
[project]
name = "toolhive-doc-mcp"
version = "1.0.0"
description = "MCP server for semantic search over Stacklok documentation"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=0.1.0",
    "uvicorn>=0.30.0",
    "sqlite-vec>=0.1.0",
    "openai>=1.0.0",
    "markdown-it-py>=3.0.0",
    "mdit-py-plugins>=0.4.0",
    "tiktoken>=0.7.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiohttp>=3.9.0",
    "fastembed>=0.7.3",
    "beautifulsoup4>=4.14.2",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.4",
    "ty>=0.0.1a18",
    "pytest-cov>=7.0.0",
]
security = [
    "safety>=3.2.11",
    "bandit[toml]>=1.8.0",
    "pip-audit>=2.9.0",
    "cyclonedx-bom>=6.0.0",
]

[tool.uv.sources]
toolhive-doc-mcp = { workspace = true }

[project.scripts]
toolhive-doc-mcp = "src.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"
lint.select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
lint.ignore = []

[tool.ty.src]
exclude = ["tests"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "--strict-config",
    "-v",
    "--cov=src",
    "--cov-report=term-missing",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::ResourceWarning",
]
